用户订阅迁移、用户Id调整为long

+++ KdyWeb.NetCore/KdyWeb.BaseInterface/Service/ILoginUserInfo.cs
@@ -25,6 +25,6 @@ namespace KdyWeb.BaseInterface.Service
         /// <summary>
         /// 用户Id
         /// </summary>
-        int? UserId { get; set; }
+        long? UserId { get; set; }
     }
 }
